created Setup.h Header File converted everything else to .cpp
