+++ src/Clock_Setup.cpp
@@ -1,5 +1,4 @@
-#include <Arduino.h>
-#include <wiring.h>
+#include "Setup.h"
 
 void Clock_Setup() {
     //more to come
